requirements.txt
djangorestframework
markdown
django-filter
asgiref
Django
pytz
sqlparse
drf-flex-fields
drf-extensions
pyyaml
uritemplate
